Working class fields for the instance object.
self.all_devices now represent all devices found in the domain.
self.active_devices represent all devices found to be active within the given timeframe in the domain.
self.inactive_devices represent all devices found to be inactive, that is, unused, within the given
timeframe in the domain. This is simply a subtraction of the active devices from all devices.

Instance variables 'deprovisioned', 'disabled', 'provisioned' represent devices in the inactive
devices list that comply with the corresponding status.

These will later be added to one separate workbook in an excel file for each one.