Revoked features: - Default user switch - get_user_id() method
Since an update with GAM and the fork of gam known as GAMADV-XTD uses another authentication
method with GAM -> Google, the way of parsing a .txt file in to .json is deprecated.

This means the user has to manually enter their username to tell Oblivio where to upload the
file.